warning in trap.c that recently appeared
src/trap.c(5410): warning C4389: '==': signed/unsigned mismatch

(cherry picked from commit 420863252ace6aaa3a109bd0c35715f15c508a2a)

+++ src/trap.c
@@ -5831,7 +5831,7 @@ count_traps(int ttyp)
     struct trap *trap = g.ftrap;
 
     while (trap) {
-        if (trap->ttyp == ttyp)
+        if ((int) trap->ttyp == ttyp)
             ret++;
         trap = trap->ntrap;
     }
